+++ Code/Commando/dlgcncpurchasemenu.cpp
@@ -154,7 +154,7 @@ CNCPurchaseMenuClass::On_Init_Dialog (void)
 				//
 				//	Configure the merchandise settings
 				//
-				int cost = Definition->Get_Cost (index) * CostScalingFactor;
+				int cost = int(Definition->Get_Cost (index) * CostScalingFactor);
 				ctrl->Set_Text (Definition->Get_Name (index));
 				ctrl->Set_Cost (cost);
 				ctrl->Set_Texture (Definition->Get_Texture (index));
@@ -326,7 +326,7 @@ CNCPurchaseMenuClass::Add_Item_To_Shopping_Cart (int ctrl_id)
 	WWASSERT(COMBAT_STAR->Get_Player_Data() != NULL);
 
 	int funds	= (int) COMBAT_STAR->Get_Player_Data ()->Get_Money ();
-	int cost		= Definition->Get_Cost (item_index) * CostScalingFactor;
+	int cost		= int(Definition->Get_Cost (item_index) * CostScalingFactor);
 	if ((TotalCost + cost) <= funds) {
 
 		//
@@ -848,7 +848,7 @@ CNCPurchaseMenuClass::Update_Enabled_Status (void)
 				//
 				//	Configure the merchandise settings
 				//
-				int cost = Definition->Get_Cost (index) * CostScalingFactor;
+				int cost = int(Definition->Get_Cost (index) * CostScalingFactor);
 
 				//
 				//	Disable any options that cost money if production is disabled
